Brazil-E-commerce-site/airflow/dags/scripts/requirements.txt
boto3
python-dotenv
findspark
pyspark